% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_fiscal_quarter.R
\name{ks_fiscal_quarter}
\alias{ks_fiscal_quarter}
\title{A period Function}
\usage{
ks_fiscal_quarter(x, with_year = FALSE)
}
\arguments{
\item{x}{Do you love cats? Defaults to TRUE.}

\item{with_year}{Do you love cats? Defaults to FALSE.}
}
\description{
This function allows you to turn any date into a fiscal quarter for ks.
}
\examples{
ks_fiscal_quarter()
}
\keyword{fiscal}
\keyword{quarter}

