% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_fiscal_year.R
\name{ks_fiscal_year}
\alias{ks_fiscal_year}
\title{A period Function}
\usage{
ks_fiscal_year(x)
}
\arguments{
\item{x}{A date object.}
}
\description{
This function allows you to turn any date into a fiscal quarter for ks.
}
\examples{
ks_fiscal_year()
}
\keyword{fiscal}
\keyword{quarter}

